Changed everything to OOP
